Further splitup of EditScreen into its components
